changed component value datatype to double, added converting to rot and pot, master can send boundsinfo to a node and a node can receive it and assign the bounds to a component
Problem with packet ids as I am sending a struct for the bounds which doesnt have a id... Now the receiver function is just an if else id == reset and else if the struct. The problem is that you can still read the id from the struct which can give some unwanted results. I will fix this next.